MLA/requirements.txt
torch>=2.0.0
transformers>=4.35.0
transformers_stream_generator
tiktoken
transformer_lens
einops
jaxtyping

scikit-learn
datasets
pandas
numpy

tqdm
colorama
requests

huggingface_hub

python-dotenv